Refactor endpoint trace graph duration helper
